Created file to handle writing to std
