+++ src/commands/echo_utils.c
@@ -37,20 +37,6 @@ t_bool	parse_n_flag(t_arg *input)
 	return (FALSE);
 }
 
-void	write_to_stdout(const char *string_to_write)
-{
-	const int	len_inside = ft_strlen(string_to_write);
-
-	write(STDOUT_FILENO, string_to_write, len_inside);
-}
-
-void	write_to_stderr(const char *string_to_write)
-{
-	const int	len_inside = ft_strlen(string_to_write);
-
-	write(STDERR_FILENO, string_to_write, len_inside);
-}
-
 void	trim_extra_spaces_between_words(t_arg *arg, t_buffer *buffer)
 {
 	if (isspace(*arg->start) && *(arg->start + 1))
