+++ src/commands/echo_utils.h
@@ -5,8 +5,6 @@
 # include <commands/buffer.h>
 
 t_bool	parse_n_flag(t_arg *input);
-void	write_to_stdout(const char *string_to_write);
-void	write_to_stderr(const char *string_to_write);
 void	trim_extra_spaces_between_words(t_arg *arg, t_buffer *buffer);
 
 #endif
