+++ src/executor/executor_utils.h
@@ -5,7 +5,7 @@
 # include <executor/redirection.h>
 
 void	wait_for_all_processes(int num_of_processes);
-void	handle_errors(int error_code, char *location);
+void	handle_errors(int error_code, const char *location);
 int		execute_command(const char *path, char *argv[], char *env[]);
 void	execute_system_command(const t_command *command, char *env[]);
 int		create_new_process(t_multi_pipes *pipes);
