+++ src/executor/run_commands.h
@@ -11,7 +11,6 @@
 int		run_pipes(const char *env[]);
 int		run_commands(t_command commands[],
 			int num_of_commands, char *env[]);
-void	handle_errors(int error_code, char *location);
 void	create_table(t_command commands[], char *arg, char *path);
 
 #endif
